Initial commit
A basic layout with a header, a footer and a navigation sidebar on the
right, using HTML5 and CSS.

The "Blog" sidebar section entries will eventually be replaced with an
list of links to the most recent blog entries by Jekyll.  The "Latest
Version" and "Older Versions" sidebar sections will also be automated in
time, assuming there's a good way of doing so.  The rest of the sidebar
should be more-or-less static.

Interesting note: The papyrus.jpg background image here was used on a
few pages in the 3.3.0 copy of the website, but links to it were broken
when the website was updated for newer versions of NetHack, so I decided
to bring it back so people wouldn't have to stare at a bright white
page.