Split out layout, add "About NetHack" page
index.html is currently a stub; it will eventually serve as a 'landing
page' to give visitors a quick impression of what NetHack is, how it
plays and how they can play it themselves.

The contents of the "About NetHack" page are extracted from the Basic
Information section of the current site's info page.  `about.md` is
converted to an HTML page named `about.html` by Jekyll.