Highlight and delink current page in nav sidebar
This gives visitors a sense of where they are on the website by looking
at the navigation sidebar.

If there's a way to fetch a page by its URL and access its title, it may
be possible to improve this further by using that by default instead of
stating it redundantly with the include tags.