About NetHack: Expand this page a lot
This is the first batch of truly original text content for this website,
as opposed to aping off of the existing website.  It describes how
NetHack is a roguelike, and how it differs from other roguelikes.

I've gone with bold and italics for almost every mention of a proper
noun, partly as a typographic convention, but mostly to visually break
up long passages of text to make them easier to read.  This does however
mean that every mention of "NetHack" is surrounded by three asterisks,
which is how it's done in Markdown.

Also, style `code` tags a bit since they're used to denote some example
commands near the end of the page.