Update blog nav sidebar subsection for April 3, 2015 announcement
The existing www.nethack.org website was updated with an announcement of
continued development, so I may as well update the fake blog section.

Reminder that this navigation sidebar subsection will eventually be
generated automatically, so this kind of manual updating won't be needed
for the final version of this website.