Add "Play Online" section to navigation sidebar
A number of online servers exist for NetHack and NetHack variants, but
it's not obvious how to play on them, why it's worth playing on them or
even that they exist, so give them a bit of exposure.

"Playing in Browser" would be a page linking to ways that various
NetHack servers can be played online without downloading any special
client software, such as:

* nethack.alt.org - Flash and Java clients.  Hosts its own patched
  version of NetHack.

* webhack.webhook.org - Web terminal: needs a browser with websocket
  support.  Connects to nethack.alt.org but is otherwise unaffiliated
  with it.

* nethack.xd.cm - Had a web terminal back when it was acehack.de, but
  currently doesn't; I think getting it back up on the new server is in
  progress.  Hosts NAOHack along with a few NetHack variants.

* Ebonhack and projects related to it may still be around, and if they
  are, they're worth mentioning somewhere in this section of the site.

All of this is subject to change, like all parts of the sidebar.