Create and fill in Frequenty Asked Questions
This FAQ page also gives a nice demo of the table-of-contents feature,
which isn't standard Markdown syntax, but rather is specific to
kramdown, the Markdown engine used by Jekyll by default.

Most of this was ported from the Information page of the existing
NetHack website.  A *lot* of this needs to be reviewed; I have left
comments throughout the page for things that stuck out to me.

A section for really basic gameplay questions is also conspicuously
absent.