Add Basic Gameplay section to FAQ
This is filled with all of the most common issues that confront players
in their first hour or so of playing NetHack.

All of the info here is technically in the Guidebook as well, but asking
players starting out to read a 50-odd page manual for basic gameplay
info is probably going to turn them away from playing altogether.