Remove Android port from "Things we are NOT doing" in the FAQ
The dev team should keep this avenue open for themselves, in case they
want to take on an Android developer in the future.

Android as a platform has broad reach, freely available dev tools and
people that already know how to adapt NetHack to work on it, so the only
thing the dev team would need to do to support it is ask for help.