Make a start on the blog
Ported a whole bunch of the news on the existing website to test it.
There are a lot of links in them that still need to be filled in.

The blog sub-section of the navigation sidebar is now functional.  The
author for all of the posts have been set to 'keni' which I presume is
the online handle for Kenneth Lorber, the current maintainer of the
website.

The blog itself consists of the index page that lists all the posts, and
individual pages for each of the posts themselves.  There are unfinished
links for the post authors and tags, features that have not been
implemented yet.