Add author pages to blog
Blog posts can now be browsed by authors, each of which gets an
individual page listing all of their blog posts, plus a short blurb.

There's also a master author list that links to each of these author
pages, along with the number of posts that they've contributed to the
blog.

To add an author, just add a file with the author name under the
`_authors` directory, with contents similar to the one in
`_authors/keni.md` (name, title, Markdown-formatted blurb).  The reason
for the redundant specifications of `name` and `title` is due to a
limitation in how Jekyll works, i.e. the data available to collection
entries when output to a single page versus via its collection are
different.