Add page for NetHack References
These are all the quotes and links from the existing NetHack website's
front page, given a page all to themselves.  A snippet of one of the
references here would go well with the welcome page of this website.

*Lots* of links here need to be checked; those that are broken need to
be pointed to their current locations if they can be found, or the Web
Archive otherwise.  The perils of deep links...

I decided that "NetHack meets T-rex" USENET post did not deserve its own
page like it has on the existing NetHack website, so it's been placed at
the end of this one instead.