Turn quotes into site data, show one on the welcome page
New quotes can be added by putting them at the top of
`_data/quotes.yml`; the contents and source are both interpreted as
Markdown.

Using Jekyll data for this might be a bit overkill, but I really want
there to be as little redundancy as possible, because redundancy means
things go out-of-date and thus get updated less often.