Link to "Freenode" on NetHack Wiki for "IRC Channels" in nav sidebar
I originally had a full page typed up for NetHack-related IRC channels
on this website proper, but then I realized that NetHack Wiki had a page
on it already.

It's better for the link to lead to the wiki, since then the community
can keep the info up-to-date instead of leaving it to the maintainer of
this website.