Add Legacy Bug Search link to Development nav sidebar section
The existing www.nethack.org website has as Perl CGI script that lets
you search through bug descriptions.  Ideally, this should occur through
a proper public-facing bug tracker instead, such as trac or Bugzilla,
but there a lot of bugs recorded in there, so it's best to have the Perl
script available until all the bugs are migrated to a proper bug
tracking system and the system goes live.