Add "Download it now!" link to latest version release date box
This is ONLY shown for the most recent version (i.e. the top-most entry
of `_data/versions.yml`) and not just any current version.  It's a prose
cue for people who don't see the sub-navigation bar, and is a tiny bit
more incentive for people to get the *latest* version, not just any
"green date box" version.