Reduce site generation time to a third of what it was
Before: 29 seconds; After: 11 seconds.

The titles of navigation links are now pulled from `_data/nav.yml`
instead of from the front matter of the page itself, which unfortunately
means that titles of pages that appear in navigation contexts need to be
typed twice (once here and once in the page's front matter).