Add NetHack 3.3.1 Downloads page
The legacy downloads pages for individual platforms still need to be
ported, since they're linked from this page.

I'm assuming that many of the platforms no longer have active downloads
since the 3.3.1 downloads directory in the SourceForge files area
doesn't have anything more than what is linked on this downloads page,
but I'll need to check around to be sure.