Make blog post dates stand out more
Dates for blog posts are now bordered by a small box, and the date
itself appears in bold.  This makes it easier to skim over the dates of
blog posts when they appear in a list, and make it stick out more for
the latest post excerpt on the welcome page.