Add new page: Using Git with NetHack
This provides a guide to using Git to create a fork of NetHack,
preparing changes in a way that is suitable for upstream inclusion and
keeping the fork up-to-date with ongoing upstream development.

Creating a fork and sending pull requests assume that things will be
done via GitHub, otherwise those steps will be a lot more complicated
and error-prone.

This whole thing still needs to be proof-read.