NetHack 3.6.0 released
Took me about 45 minutes to put all this together, which is a bit longer
than I expected given all the instrumentation to help with updating the
site for a new release.

I didn't have to touch a single page that wasn't related to 3.6.0 to do
this update, which proves that the concept of this website works.